+++ components/bottombar/api/src/main/java/com/flipperdevices/bottombar/api/BottomNavigationApi.kt
@@ -1,10 +1,11 @@
 package com.flipperdevices.bottombar.api
 
+import com.flipperdevices.deeplink.model.Deeplink
 import com.github.terrakok.cicerone.androidx.FragmentScreen
 
 /**
  * Class which provide api to bottombar module
  */
 interface BottomNavigationApi {
-    fun getBottomNavigationFragment(): FragmentScreen
+    fun getBottomNavigationFragment(deeplink: Deeplink? = null): FragmentScreen
 }